move aggregation constants to aggregations.js
