+++ src/internal-packages/chart/lib/constants/index.js
@@ -1,7 +1,8 @@
+const aggregations = require('./aggregations');
 const channels = require('./channels');
 const reductions = require('./reductions');
 const spec = require('./spec');
 const types = require('./types');
 
 module.exports = Object.freeze(Object.assign({},
-  channels, reductions, spec, types));
+  aggregations, channels, reductions, spec, types));
